pyproject.toml
[tool.poetry]
name = "custom-agency-swarm"
version = "0.0.0"
description = "swarm play"
authors = ["VRSEN"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.0"
openai = "^1.3.6"
instructor = "0.3.4"
deepdiff = "6.7.1"
termcolor = "2.3.0"
python-dotenv = "1.0.0"
langchain = "^0.0.345"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
jupytext = "^1.15.2"
pre-commit = "^3.5.0"
gradio = "^4.7.1"
requests-oauthlib = "^1.3.1"
python-upwork-oauth2 = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
